requirements.txt
cryptography==43.0.1
Django==4.2.15
django-redis==5.3.0
requests==2.32.2
gunicorn==22.0.0
PyMySQL==1.1.1
matplotlib==3.5.3
numpy==1.24.4
pandas==1.1.5
psycopg2-binary==2.9.1
pyodbc==4.0.35
mariadb==1.1.10
redis==4.6.0
SQLAlchemy==1.4.46
sqlparse==0.5.0
typing-extensions==4.7.1
django-ace==1.23.4
celery==5.3.1
gunicorn==22.0.0
django-recaptcha==3.0.0
Jinja2==3.1.4
djangorestframework==3.15.2
django-simple-history==3.4.0
django-guardian==2.4.0